# docs/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RecoEngine API",
    "description": "RecoEngine turns any catalogue (jobs, dishes, products, courses, or your own items) into\nrecommendations. Upload items, and they are embedded with OpenAI and stored in a vector\nindex; then ask for items similar to a text, an existing item, or a profile.\n\n**Authentication:** send your API key in the `X-API-Key` header. Keys are created at\nregistration and on the API Keys page of the dashboard.\n\n**Errors** always look like `{\"error\": {\"code\": \"...\", \"message\": \"...\", \"details\": [...]}}`.\n`429` and `503` responses include a `Retry-After` header (seconds).\n\n**Rate limits:** 100 requests per minute per API key, 10,000 uploaded items per tenant per\nUTC day.",
    "version": "0.1.0",
    "contact": {
      "name": "RecoEngine",
      "url": "https://recoengine.io"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.recoengine.io",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/tenants": {
      "post": {
        "tags": [
          "Tenants"
        ],
        "summary": "Register a new tenant",
        "operationId": "createTenant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin API disabled (no ADMIN_API_KEY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminKeyAuth": []
          }
        ],
        "description": "Operator only: create a tenant without a password. Customers sign up with `/auth/register`."
      }
    },
    "/api/v1/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Get tenant details",
        "operationId": "getTenant",
        "security": [
          {
            "AdminKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin API disabled (no ADMIN_API_KEY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Operator only: a tenant by id."
      }
    },
    "/api/v1/tenants/{tenant_id}/api-keys": {
      "post": {
        "tags": [
          "Tenants"
        ],
        "summary": "Generate a new API key (the plain key is returned only once)",
        "operationId": "createTenantApiKey",
        "security": [
          {
            "AdminKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreatedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Tenant is inactive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Operator only: create an API key for a tenant, e.g. for support or migrations."
      },
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "List API keys (prefix only)",
        "operationId": "listTenantApiKeys",
        "security": [
          {
            "AdminKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  },
                  "title": "Response List Api Keys Api V1 Tenants  Tenant Id  Api Keys Get"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin API disabled (no ADMIN_API_KEY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Operator only: API keys of a tenant by id."
      }
    },
    "/api/v1/tenants/{tenant_id}/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "Tenants"
        ],
        "summary": "Revoke an API key",
        "operationId": "revokeTenantApiKey",
        "security": [
          {
            "AdminKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin API disabled (no ADMIN_API_KEY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Operator only: revoke an API key of a tenant, e.g. a leaked one."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Human-readable label, e.g. 'production-backend'."
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "Optional expiry. The key is rejected after this time."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ApiKeyCreate"
      },
      "ApiKeyCreatedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Id",
            "description": "The user who created the key, if they still exist."
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "The plain API key. Returned only once, at creation."
          },
          "warning": {
            "type": "string",
            "title": "Warning",
            "default": "Save this key, it won't be shown again"
          },
          "display_key": {
            "type": "string",
            "title": "Display Key",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key_prefix",
          "is_active",
          "last_used_at",
          "expires_at",
          "created_at",
          "api_key",
          "display_key"
        ],
        "title": "ApiKeyCreatedResponse"
      },
      "ApiKeyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Id",
            "description": "The user who created the key, if they still exist."
          },
          "display_key": {
            "type": "string",
            "title": "Display Key",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key_prefix",
          "is_active",
          "last_used_at",
          "expires_at",
          "created_at",
          "display_key"
        ],
        "title": "ApiKeyResponse",
        "description": "Safe representation of a key: never includes the plain key or its hash."
      },
...
      "DomainConfig": {
        "properties": {
          "primary_embedding_field": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,63}$",
            "title": "Primary Embedding Field"
          },
          "searchable_fields": {
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,63}$"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Searchable Fields"
          },
          "filter_fields": {
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,63}$"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Filter Fields"
          },
          "item_label": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Item Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "primary_embedding_field",
          "searchable_fields",
          "item_label"
        ],
        "title": "DomainConfig",
        "description": "Describes a tenant's item schema. This is what makes the engine domain-agnostic.",
        "example": {
          "filter_fields": [
            "location",
            "category"
          ],
          "item_label": "job",
          "primary_embedding_field": "description",
          "searchable_fields": [
            "title",
            "description",
            "tags"
          ]
        }
      },
...
      "DomainType": {
        "type": "string",
        "enum": [
          "HR",
          "FOOD",
          "ECOMMERCE",
          "EDTECH",
          "CUSTOM"
        ],
        "title": "DomainType"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "TenantCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "domain_type": {
            "$ref": "#/components/schemas/DomainType"
          },
          "domain_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DomainConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional for HR/FOOD/ECOMMERCE/EDTECH (a preset is used); required for CUSTOM."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "email",
          "domain_type"
        ],
        "title": "TenantCreate"
      },
      "TenantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "domain_type": {
            "$ref": "#/components/schemas/DomainType"
          },
          "domain_config": {
            "$ref": "#/components/schemas/DomainConfig"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "domain_type",
          "domain_config",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "TenantResponse"
      },
...
    },
    "securitySchemes": {
      "AdminKeyAuth": {
        "type": "apiKey",
        "description": "Operator key (ADMIN_API_KEY). Never give it to tenants.",
        "in": "header",
        "name": "X-Admin-Key"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "Your API key, e.g. `reco_…`. Create keys in the dashboard or via /me/api-keys.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}